+++ arrayjit/lib/utils.ml
@@ -118,6 +118,7 @@ let known_config_keys =
       (* Low-level / optimization *)
       "virtualize_max_visits";
       "virtualize_max_inline_reduction";
+      "virtualize_max_inline_fanin";
       "enable_device_only";
       "inline_scalar_constexprs";
       "inline_simple_computations";
@@ -216,6 +217,7 @@ let config_key_classification : (config_key_class * string * string list) list =
       [
         "virtualize_max_visits";
         "virtualize_max_inline_reduction";
+        "virtualize_max_inline_fanin";
         "enable_device_only";
         "inline_scalar_constexprs";
         "inline_simple_computations";